+++ src/block_splitting.jl
@@ -3,7 +3,7 @@ mutable struct Block # MCU
 end
 
 function split_into_blocks(x::Matrix{Float32}; blocksize=8)::Matrix{Block}
-    blocks::Matrix{Block} = map(x->Block(Matrix{Float32}(undef, (blocksize,blocksize))), zeros(Int64.(size(x)./blocksize)...))
+    blocks::Matrix{Block} = map(x->Block(Matrix{Float32}(undef, (blocksize,blocksize))), zeros(Int.(size(x)./blocksize)...))
     for (idxi,i)=enumerate(1:blocksize:size(x)[1]), (idxj,j)=enumerate(1:blocksize:size(x)[2])
         block = blocks[idxi,idxj]
         for (idxii,ii)=enumerate(i:i+(blocksize-1)), (idxjj,jj)=enumerate(j:j+(blocksize-1))
@@ -14,7 +14,7 @@ function split_into_blocks(x::Matrix{Float32}; blocksize=8)::Matrix{Block}
 end
 
 function join_blocks(x::Matrix{Block}; blocksize=8)::Matrix{Float32}
-    retimg = Matrix{Float32}(undef, Int64.(size(x).*blocksize))
+    retimg = Matrix{Float32}(undef, Int.(size(x).*blocksize))
     for (idxi,i)=enumerate(1:blocksize:size(retimg)[1]), (idxj,j)=enumerate(1:blocksize:size(retimg)[2])
         block = x[idxi,idxj]
         for (idxii,ii)=enumerate(i:i+blocksize-1), (idxjj,jj)=enumerate(j:j+blocksize-1)
