+++ src/downsampling.jl
@@ -2,8 +2,8 @@ using Statistics
 
 #lossy
 function downsample_4_2_0(x::Matrix{Float32})::Matrix{Float32}
-    retimg = Matrix{Float32}(undef, Int64.(size(x)./2))
-    for i=1:Int64(size(x)[1]/2), j=1:Int64(size(x)[2]/2)
+    retimg = Matrix{Float32}(undef, Int.(size(x)./2))
+    for i=1:Int(size(x)[1]/2), j=1:Int(size(x)[2]/2)
         block_lower = 2*i
         block_upper = block_lower-1
         block_right = 2*j
@@ -16,7 +16,7 @@ end
 
 function upsample_4_2_0(x::Matrix{Float32})::Matrix{Float32}
     retimg = Matrix{Float32}(undef, (size(x).*2))
-    for i=1:Int64(size(x)[1]), j=1:Int64(size(x)[2])
+    for i=1:Int(size(x)[1]), j=1:Int(size(x)[2])
         block_lower = 2*i
         block_upper = block_lower-1
         block_right = 2*j
